+++ src/core/MCore.c
@@ -96,3 +96,45 @@ int NodeProcess(module *m, pool **all, pool *current) {
   return mstat;
 }
 
+/**
+ * @brief The Loop prepare hook
+ *
+ * @param m The module pointer
+ * @param all The pointer to the all pools
+ * @param current The current pool pointer
+ *
+ * @return 0 on success, error code otherwise
+ */
+int LoopPrepare(module *m, pool **all, pool *current) {
+  int mstat = SUCCESS;
+  setup *s = &(m->layer.setup);
+  query *q;
+
+  q = LoadSym(m, "LoopPrepare", LOAD_DEFAULT);
+  if (q) mstat = q(m->mpi_size, m->node, all, current, s);
+  CheckStatus(mstat);
+
+  return mstat;
+}
+
+/**
+ * @brief The Loop process hook
+ *
+ * @param m The module pointer
+ * @param all The pointer to the all pools
+ * @param current The current pool pointer
+ *
+ * @return 0 on success, error code otherwise
+ */
+int LoopProcess(module *m, pool **all, pool *current) {
+  int mstat = SUCCESS;
+  setup *s = &(m->layer.setup);
+  query *q;
+
+  q = LoadSym(m, "LoopProcess", LOAD_DEFAULT);
+  if (q) mstat = q(m->mpi_size, m->node, all, current, s);
+  CheckStatus(mstat);
+
+  return mstat;
+}
+